#Chore Fix cypress db initializer

+++ identityprovider/src/main/resources/database/data-cypress.sql
@@ -1,8 +1,8 @@
 -- Delete any cypress test users
-DELETE FROM user_roles WHERE user_user_id IN (SELECT user_id FROM user WHERE username LIKE 'Cypress%');
-DELETE FROM users_groups WHERE group_id IN (SELECT group_id FROM groups WHERE long_name LIKE '%Cypress%');
+DELETE FROM user_roles WHERE user_user_id IN (SELECT user_id FROM user WHERE username LIKE '%Cypress%');
+DELETE FROM users_groups WHERE group_id IN (SELECT group_id FROM groups WHERE long_name LIKE '%Cypress%') OR group_id IN (SELECT user_id FROM user WHERE username LIKE '%Cypress%');
 DELETE FROM groups WHERE long_name LIKE '%Cypress%';
-DELETE FROM user WHERE username LIKE 'Cypress%';
+DELETE FROM user WHERE username LIKE '%Cypress%';
 
 -- Create admin user, large user id to reduce change of overlap
 INSERT IGNORE INTO user (user_id, bio,date_created,email,first_name,last_name,nickname,password,profile_image_path,pronouns,salt,username) VALUES (1285321, 'Automated test admin','2022-07-15 21:45:57.287000000','automated@cypress.com','Cypress','Admin','','JRqHHIu2hllQo8uswun8chNpkSnX2ERuhyhToKmSdYc=',null,'','aNAKNt2LHIfchQoLQ0Aryg==','CypressAdmin');
